% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoPX_ComputeR.R
\name{CoPX_ComputeR}
\alias{CoPX_ComputeR}
\title{CoPX_ComputeR}
\usage{
CoPX_ComputeR(My, Fz)
}
\arguments{
\item{My}{numeric vector specifying the moment on the Y-axis}

\item{Fz}{numeric vector specifying the force on the Z-axis}
}
\value{
A numeric vector containing the Center of pressure values (cm) on the X-axis
}
\description{
A function to compute Center of pressure (CoP) values on the medio-lateral axis (X-axis). The function only requires moments of forces on the Y-axis and Force on the Z-axis which are provided in most postural tools outputs.
}
\examples{
path_to_data <- system.file("extdata", package = "BalanceMate")
Data <- Merge_PosData(path_to_data, SampleRate = 100, SessionDuration = 331)

CoPX_ComputeR(Data$My, Data$Fz)
}
\seealso{
\code{\link[=CoPY_ComputeR]{CoPY_ComputeR()}} for computing CoP-Y values
}
