% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoPY_ComputeR.R
\name{CoPY_ComputeR}
\alias{CoPY_ComputeR}
\title{CoPY_ComputeR}
\usage{
CoPY_ComputeR(Mx, Fz)
}
\arguments{
\item{Mx}{numeric vector specifying the moment on the Y-axis}

\item{Fz}{numeric vector specifying the force on the Z-axis}
}
\value{
A numeric vector containing the Center of pressure values (cm) on the Y-axis
}
\description{
A function to compute Center of pressure (CoP) values on the antero-posterioir axis (Y-axis), a.k.a. CoP-Y. The function only requires moments of forces on the X-axis and Force on the Z-axis which are provided in most postural tools outputs.
}
\examples{
path_to_data <- system.file("extdata", package = "BalanceMate")
Data <- Merge_PosData(path_to_data, SampleRate = 100, SessionDuration = 100)

CoPY_ComputeR(Data$Mx, Data$Fz)
}
\seealso{
\code{\link[=CoPX_ComputeR]{CoPX_ComputeR()}} for computing CoP-X values
}
