% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{Postural_DataB_txt}
\alias{Postural_DataB_txt}
\title{Postural data - Block B - txt file}
\format{
\subsection{\code{Postural_DataB}}{

A data frame with 33,100 rows and 6 columns:
\describe{
\item{Mx}{Un-named column corresponding to Mx}
\item{My}{Un-named column corresponding to My}
\item{Mz}{Un-named column corresponding to Mz}
\item{Fx}{Un-named column corresponding to Fx}
\item{Fy}{Un-named column corresponding to Fy}
\item{Fz}{Un-named column corresponding to Fz}
}
}
}
\source{
<Monéger, J., Besche-Richard, C., Chaby, L., & Vergilino-Perez D. (Submitted in Emotion). Should I Stay or Should I go: An Exploration of Spontaneous Postural Behaviours Following Threatening Emotion Perception.>
}
\usage{
```r
# To load the data from the text file:
path_to_data <- system.file("extdata", "Postural_DataB.txt", package = "BalanceMate")
Postural_DataB_txt <- read.table(path_to_data, header = TRUE, sep = "\t")
```
}
\description{
A .txt file version of a full block of postural measurement during a session in the Monéger et al. (submitted) experiment. Block duration is 331s, and Sample rate was 100Hz. As such, each block corresponds to a data frame of 33,100 rows with the 6 exported columns from AMTI NetForce, unconveniently un-named. They correspond to Mx, My, Mz, Fx, Fy, Fz.
This is a plain text version of the postural data for Block B. It is stored as a .txt file in the \code{inst/extdata} directory of the BalanceMate package.
}
\examples{
path_to_data <- system.file("extdata", "Postural_DataB.txt", package = "BalanceMate")
Postural_DataB <- read.table(path_to_data, header = TRUE, sep = "\t")

}
