% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{Postural_DataD}
\alias{Postural_DataD}
\title{Postural data - Block D (R data)}
\format{
\subsection{\code{Postural_DataD}}{

A data frame with 33,100 rows and 6 columns:
\describe{
\item{Mx}{Un-named column corresponding to Mx}
\item{My}{Un-named column corresponding to My}
\item{Mz}{Un-named column corresponding to Mz}
\item{Fx}{Un-named column corresponding to Fx}
\item{Fy}{Un-named column corresponding to Fy}
\item{Fz}{Un-named column corresponding to Fz}
}
}
}
\source{
<Monéger, J., Besche-Richard, C., Chaby, L., & Vergilino-Perez D. (Submitted in Emotion). Should I Stay or Should I go: An Exploration of Spontaneous Postural Behaviours Following Threatening Emotion Perception.>
}
\usage{
data(Postural_DataD)
}
\description{
A full block of postural measurement during a session in the Monéger et al. (submitted) experiment. Block duration is 331s, and Sample rate was 100Hz. As such, each block corresponds to a data frame of 33,100 rows with the 6 exported columns from AMTI NetForce, unconveniently un-named. They correspond to Mx, My, Mz, Fx, Fy, Fz.
}
\keyword{datasets}
