% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{Postural_DataF_txt}
\alias{Postural_DataF_txt}
\title{Postural data - Block F - txt file}
\format{
\subsection{\code{Postural_DataF}}{

A data frame with 33,100 rows and 6 columns:
\describe{
\item{Mx}{Un-named column corresponding to Mx- moment around the X axis}
\item{My}{Un-named column corresponding to My- moment around the Y axis}
\item{Mz}{Un-named column corresponding to Mz- moment around the Z axis}
\item{Fx}{Un-named column corresponding to Fx- force in the X axis}
\item{Fy}{Un-named column corresponding to Fy- force in the Y axis}
\item{Fz}{Un-named column corresponding to Fz- force in the Z axis}
}
}
}
\source{
<Monéger, J., Besche-Richard, C., Chaby, L., & Vergilino-Perez D. (Submitted in Emotion). Should I Stay or Should I go: An Exploration of Spontaneous Postural Behaviours Following Threatening Emotion Perception.>
}
\usage{
# Postural_DataF_txt
}
\description{
A .txt file version of a full block of postural measurement during a session in the Monéger et al. (submitted) experiment. Block duration is 331s, and Sample rate was 100Hz. As such, each block corresponds to a data frame of 33,100 rows with the 6 exported columns from AMTI NetForce, unconveniently un-named. They correspond to Mx, My, Mz, Fx, Fy, Fz.
This is a plain text version of the postural data for Block F. It is stored as a .txt file in the \code{inst/extdata} directory of the BalanceMate package.
}
\examples{
path_to_data <- system.file("extdata", "Postural_DataF.txt", package = "BalanceMate")
Postural_DataF <- read.table(path_to_data, header = TRUE, sep = "\t")

}
