% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spaghett.R
\name{Spaghett}
\alias{Spaghett}
\title{Create a "spaghetti plot" from your postural data.}
\usage{
Spaghett(
  df,
  time_col = "Time",
  copx_col = "CoP_X",
  copy_col = "CoP_Y",
  Title = "Spaghetti plot",
  xlab = "Mean CoP_X",
  ylab = "Mean CoP_Y"
)
}
\arguments{
\item{df}{dataframe containing the data you want to plot}

\item{time_col}{column from dataframe that indicates the time course of postural measurements}

\item{copx_col}{column from dataframe that contains the CoP-X displacements}

\item{copy_col}{column from dataframe that contains the CoP-Y displacements}

\item{Title}{Character string specifying a title for your plot (optionnal)}

\item{xlab}{Character string specifying the x-axis name (optional)}

\item{ylab}{Character string specifying the y-axis name (optional)}
}
\value{
a graphical object
}
\description{
This function creates a spaghetti plot - a plot that illustrates the CoP displacements stored in a data frame. The data frame should contain minimally a column storing the time course, a column storing CoP-X values, and a column storing CoP-Y values.
}
\examples{
data(Postural_DataA) # Load a session to plot
Spaghett(Postural_DataA, time_col = "Time", copx_col = "CoP_X", copy_col = "CoP_Y")


}
