% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Time_StampeR.R
\name{Time_StampeR}
\alias{Time_StampeR}
\title{Annotate your Postural file to describe important periods}
\usage{
Time_StampeR(
  df,
  id_col,
  sample_rate,
  protocol_duration,
  cuts = NULL,
  period_names = NULL
)
}
\arguments{
\item{df}{a data frame containing your postural data}

\item{id_col}{a column from the data frame that store the participants' unique identifiers}

\item{sample_rate}{the sample rate used in your protocol}

\item{protocol_duration}{the duration of your protocols}

\item{cuts}{a list of numeric values that indicate where your protocol should be split to distinguish between two different periods.}

\item{period_names}{a list of character strings that will be used as labels for your split. There should be one more label than cuts.}
}
\value{
a data frame with an additional column labelling each time measurement.
}
\description{
This function allows user to annotate (i.e., add a column to the data frame) a dataframe to label important periods in the session. For instance, if your protocol included training (10s) and trial (30s), you can use the \code{Time_StampeR} to add a column that will annotate the training measures and the trial measures.
}
\examples{
# Find subdirectory of Example data in the original .txt format exported from AMTI Netforce software
# Note: we need to convert compressed rdata to original txt file
files <- paste0("Postural_Data", LETTERS[1:6])

# Locate the directory containing the .Rdata files within the package
data_dir <- system.file("data", package = "BalanceMate")

# Create a temporary directory to store the .txt files
temp_data_dir <- file.path(tempdir(), "data")
dir.create(temp_data_dir, showWarnings = FALSE)

# Process each file: load, optionally add a blank row, and write to .txt
invisible(lapply(files, function(f) {
  # Load the .Rdata file from the package's extdata directory
  load(file.path(data_dir, paste0(f, ".Rdata")))

  data <- get(f)
  # Write the data to a .txt file in the temporary directory
  write.table(data, file = file.path(temp_data_dir, paste0(f, ".txt")), sep = ",",
  row.names = FALSE, col.names = FALSE, quote = FALSE)
}))

Data <- Merge_PosData(temp_data_dir, SampleRate = 100, SessionDuration = 100)

# here session: 30 s training, 3 trials (10 s fix cross + 90 s trial), 1s blank
cuts = c(30, 40, 130, 140, 230, 240, 330)
labels = c("Training",
      "FixationCross",
      "trial1",
      "FixationCross",
      "trial2",
      "FixationCross",
      "trial3",
      "blank")
Data <- Time_StampeR(df = Data,
      id_col = "file_name",
      sample_rate = 100,
      protocol_duration = 100,
      cuts = cuts,
      period_names = labels)


}
