% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EllipseArea.R
\name{compute_ellipse_area}
\alias{compute_ellipse_area}
\title{Compute X\% ellipse area (cm.cm)}
\usage{
compute_ellipse_area(
  data,
  CoPX_col,
  CoPY_col,
  ID,
  time_col = NULL,
  epoch_length = NULL,
  confint = 0.95
)
}
\arguments{
\item{data}{a data frame containing postural data (i.e., Center of pressure displacements on both the antero-posterior and medio-lateral axis)}

\item{CoPX_col}{a column from the data frame storing center of pressure displacements on the medio-lateral axis}

\item{CoPY_col}{a column from the data frame storing center of pressure displacements on the antero-posterior axis}

\item{ID}{a column from the data frame storing unique identifiers for each participant/session}

\item{time_col}{a column from the data frame storing time course over the session (optional; for computing results synthetised per epochs)}

\item{epoch_length}{an optional argument (if computing results synthetised at an epoch level) specifying length of the desired epoch}

\item{confint}{desired percentage for the confidence in the ellipse area (default is .95)}
}
\value{
a data frame with an additional 'ellipse_area' column
}
\description{
Compute X\% ellipse area (cm.cm)
}
\examples{
#Find subdirectory of Example data in the original .txt format exported from AMTI Netforce software
path_to_data <- system.file("extdata", package = "BalanceMate")
Data <- Merge_PosData(path_to_data, SampleRate = 100, SessionDuration = 100)
compute_ellipse_area(Data, "CoP_X", "CoP_Y", ID = "file_name") # Individual level

# Epochs of 1 second
compute_ellipse_area(Data, "CoP_X", "CoP_Y",
      ID = "file_name",
      time_col = "Time",
      epoch_length = 1)

# Epochs of 1 second and 80\% confidence Ellipse area
compute_ellipse_area(Data, "CoP_X", "CoP_Y",
      ID = "file_name",
      time_col = "Time",
      epoch_length = 1,
      confint = .80)

}
