% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoPY_ComputeR.R
\name{CoPY_ComputeR}
\alias{CoPY_ComputeR}
\title{CoPY_ComputeR}
\usage{
CoPY_ComputeR(Mx, Fz)
}
\arguments{
\item{Mx}{numeric vector specifying the moment on the Y-axis}

\item{Fz}{numeric vector specifying the force on the Z-axis}
}
\value{
A numeric vector containing the Center of pressure values (cm) on the Y-axis
}
\description{
A function to compute Center of pressure (CoP) values on the antero-posterioir axis (Y-axis), a.k.a. CoP-Y. The function only requires moments of forces on the X-axis and Force on the Z-axis which are provided in most postural tools outputs.
}
\examples{
# Note: we need to convert compressed rdata to original txt file
files <- paste0("Postural_Data", LETTERS[1:6])

# Locate the directory containing the .Rdata files within the package
data_dir <- system.file("data", package = "BalanceMate")

# Create a temporary directory to store the .txt files
temp_data_dir <- file.path(tempdir(), "data")
dir.create(temp_data_dir, showWarnings = FALSE)

# Process each file: load, optionally add a blank row, and write to .txt
invisible(lapply(files, function(f) {
  # Load the .Rdata file from the package's extdata directory
  load(file.path(data_dir, paste0(f, ".Rdata")))

  data <- get(f)
  # Write the data to a .txt file in the temporary directory
  write.table(data, file = file.path(temp_data_dir, paste0(f, ".txt")), sep = ",",
  row.names = FALSE, col.names = FALSE, quote = FALSE)
}))

Data <- Merge_PosData(temp_data_dir, SampleRate = 100, SessionDuration = 100)

CoPY_ComputeR(Data$Mx, Data$Fz)
}
\seealso{
\code{\link[=CoPX_ComputeR]{CoPX_ComputeR()}} for computing CoP-X values
}
