% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Merge_PosData.R
\name{Merge_PosData}
\alias{Merge_PosData}
\title{Merge and Append your Postural data}
\usage{
Merge_PosData(directory_path, SampleRate, SessionDuration)
}
\arguments{
\item{directory_path}{directory to a folder containing all the .txt outputs from AMTI Netforce that you wish to merge in a single data set.}

\item{SampleRate}{Sample rate used in your protocol}

\item{SessionDuration}{Duration of your protocol}
}
\value{
a dataframe containing all merged postural files, with additional columns: CoP-Y, CoP-X, time course, and file name.
}
\description{
This function merges all .txt files directly exported from the AMTI NetForce software. It re-names columns (un-named by the AMTI software), compute automatically CoP on both the antero-posterior (CoP-Y) and medio-lateral (CoP-X) axes, and add the time course during the session to facilitate further data management
}
\examples{
path_to_data <- system.file("extdata", package = "BalanceMate") #Find subdirectory of Example data in the original .txt format exported from AMTI Netforce software
Data <- Merge_PosData(path_to_data, SampleRate = 100, SessionDuration = 331) # Input correct arguments: in this example, the protocol was 331seconds long and the sample rate was 100Hz
}
