% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SD_CoP_ComputeR.R
\name{SD_CoP_ComputeR}
\alias{SD_CoP_ComputeR}
\title{Compute Mean Standard deviations of CoP-X and CoP-Y displacements for a specified time bin}
\usage{
SD_CoP_ComputeR(
  data,
  CoPX_col,
  CoPY_col,
  ID,
  time_col = NULL,
  epoch_length = NULL
)
}
\arguments{
\item{data}{a dataframe containing the postural data}

\item{CoPX_col}{a column in the data frame storing the CoP-X data}

\item{CoPY_col}{a column in the data frame storing the CoP-Y data}

\item{ID}{a column in the data frame storing Participants' unique identifiers}

\item{time_col}{a column in the data frame storing Time course}

\item{epoch_length}{numeric value: epoch duration over which you want to compute the sway path length (in seconds)}
}
\value{
a data frame containing the new postural measures that were synthetised at the ID level or at the epoch level (if specified).
}
\description{
Compute Mean Standard deviations of CoP-X and CoP-Y displacements for a specified time bin
}
\examples{
#Find subdirectory of Example data in the original .txt format exported from AMTI Netforce software
path_to_data <- system.file("extdata", package = "BalanceMate")
Data <- Merge_PosData(path_to_data, SampleRate = 100, SessionDuration = 331)

# Compute mean SD CoP-X and mean SD CoP-Y at the participant (ID level):
SD_CoP_ComputeR(Data, "CoP_X", "CoP_Y", "file_name")

# Synthetise SD CoP-X and SD CoP-Y for 1s time bins:
SD_CoP_ComputeR(Data, "CoP_X", "CoP_Y", "file_name", time_col = "Time", epoch_length = 1)

# Synthetise SD CoP-X and SD CoP-Y for 2s time bins:
SD_CoP_ComputeR(Data, "CoP_X", "CoP_Y", "file_name", time_col = "Time", epoch_length = 2)

}
