% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPL_ComputeR.R
\name{SPL_ComputeR}
\alias{SPL_ComputeR}
\title{Compute mean Sway Path Length}
\usage{
SPL_ComputeR(
  data,
  CoPX_col,
  CoPY_col,
  ID,
  time_col = NULL,
  epoch_length = NULL
)
}
\arguments{
\item{data}{a data frame}

\item{CoPX_col}{a column in the data frame storing the CoP-X data}

\item{CoPY_col}{a column in the data frame storing the CoP-Y data}

\item{ID}{a column in the data frame storing Participants' unique identifiers}

\item{time_col}{a column in the data frame storing Time course}

\item{epoch_length}{numeric value: epoch duration over which you want to compute the sway path length (in seconds)}
}
\value{
a synthetised data frame (i.e., shorter) with an additional column (Sway path lenght)
}
\description{
Compute mean Sway Path Length
}
\details{
If the specified epoch is not a multiple of the protocol duration (e.g., user inputs an epoch of 3s for a protocol of 10s), then the function will compute the SPL over the unbalanced time periods (in the latter example, 4 mean SPL would be computed with the three first periods corresponding to 3s intervals, but the last one corresponding to the remaining time -- that is 1s interval). A warning will be issued to inform user about unbalanced epoch computations.
}
\examples{
# Find subdirectory of Example data in the original .txt format exported from AMTI Netforce software
path_to_data <- system.file("extdata", package = "BalanceMate")
Data <- Merge_PosData(path_to_data, SampleRate = 100, SessionDuration = 100)

SPL_ComputeR(data = Data,
    CoPX_col = "CoP_X",
    CoPY_col = "CoP_Y",
    ID = "file_name",
    time_col = "Time",
    epoch_length = 1)


}
